--- Dockerfile
@@ -1,2 +1,0 @@
-FROM mongo:latest
-COPY mongo-init.js /docker-entrypoint-initdb.d/
