+++ api/src/di/dependencies.py
@@ -1,0 +1,17 @@
+from typing import Annotated
+
+from fastapi import Depends
+from reporter.service import LoLStatsService
+from pymongo import MongoClient
+
+
+def provide_mongo_client():
+    mongo_client = MongoClient()
+    
+    yield mongo_client
+    
+# Dependency function for LoLStatsService
+def get_lol_stats_service(
+    mongo_client: Annotated[MongoClient, Depends(provide_mongo_client)]
+) -> LoLStatsService:
+    return LoLStatsService(mongo_client)
