+++ api/src/main.py
@@ -1,0 +1,6 @@
+from fastapi import FastAPI
+from reporter.router import router as reporter_router
+
+app = FastAPI()
+
+app.include_router(reporter_router)
