+++ infra/mongodb/Dockerfile
@@ -1,0 +1,2 @@
+FROM mongo:latest
+COPY mongo-init.js /docker-entrypoint-initdb.d/
