+++ infra/mongodb/mongo-init.js
