--- src/di/dependencies.py
@@ -1,7 +1,0 @@
-from pymongo import MongoClient
-
-
-def provide_mongo_client():
-    mongo_client = MongoClient()
-    
-    yield mongo_client
